Create new page and moved cornhole counter to /counter
